+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # steam-family-bot-core
 
+## 0.0.6-dev.6
+
+### Patch Changes
+
+- 9418eea: generic CHANNEL_INFO update
+
 ## 0.0.6-dev.5
 
 ### Patch Changes