+++ packages/discord/CHANGELOG.md
@@ -1,5 +1,12 @@
 # steam-family-bot-discord
 
+## 0.0.6-dev.6
+
+### Patch Changes
+
+- Updated dependencies [9418eea]
+  - steam-family-bot-core@0.0.6-dev.6
+
 ## 0.0.6-dev.5
 
 ### Patch Changes