+++ packages/koishi/CHANGELOG.md
@@ -1,5 +1,13 @@
 # koishi-plugin-steam-family-bot
 
+## 0.0.6-dev.6
+
+### Patch Changes
+
+- 9418eea: generic CHANNEL_INFO update
+- Updated dependencies [9418eea]
+  - steam-family-bot-core@0.0.6-dev.6
+
 ## 0.0.6-dev.5
 
 ### Patch Changes